inst: Absolute variants are u16s

+++ src/inst.rs
@@ -88,9 +88,9 @@ pub enum ADC {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -100,9 +100,9 @@ pub enum AND {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -112,14 +112,14 @@ pub enum ASL {
 	Accumulator,
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum BIT {
 	ZeroPage(u8),
-	Absolute(u8),
+	Absolute(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -127,9 +127,9 @@ pub enum CMP {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -138,22 +138,22 @@ pub enum CMP {
 pub enum CPX {
 	Immediate(u8),
 	ZeroPage(u8),
-	Absolute(u8),
+	Absolute(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum CPY {
 	Immediate(u8),
 	ZeroPage(u8),
-	Absolute(u8),
+	Absolute(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum DEC {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -161,9 +161,9 @@ pub enum EOR {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -172,8 +172,8 @@ pub enum EOR {
 pub enum INC {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -187,11 +187,11 @@ pub enum LDA {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
-	IndirectY(u8),
+	IndirectY(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -199,8 +199,8 @@ pub enum LDX {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageY(u8),
-	Absolute(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteY(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -208,8 +208,8 @@ pub enum LDY {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -217,8 +217,8 @@ pub enum LSR {
 	Accumulator,
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -226,9 +226,9 @@ pub enum ORA {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -238,8 +238,8 @@ pub enum ROL {
 	Accumulator,
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -247,8 +247,8 @@ pub enum ROR {
 	Accumulator,
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -256,9 +256,9 @@ pub enum SBC {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -278,22 +278,22 @@ pub enum STA {
 pub enum STX {
 	ZeroPage(u8),
 	ZeroPageY(u8),
-	Absolute(u8),
+	Absolute(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum STY {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
+	Absolute(u16),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum LAX {
 	ZeroPage(u8),
 	ZeroPageY(u8),
-	Absolute(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -302,7 +302,7 @@ pub enum LAX {
 pub enum SAX {
 	ZeroPage(u8),
 	ZeroPageY(u8),
-	Absolute(u8),
+	Absolute(u16),
 	IndirectX(u8),
 }
 
@@ -310,9 +310,9 @@ pub enum SAX {
 pub enum DCP {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -321,9 +321,9 @@ pub enum DCP {
 pub enum ISC {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -332,9 +332,9 @@ pub enum ISC {
 pub enum RLA {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -343,9 +343,9 @@ pub enum RLA {
 pub enum RRA {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -354,9 +354,9 @@ pub enum RRA {
 pub enum SLO {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
@@ -365,16 +365,16 @@ pub enum SLO {
 pub enum SRE {
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 	IndirectX(u8),
 	IndirectY(u8),
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum AHX {
-	AbsoluteY(u8),
+	AbsoluteY(u16),
 	IndirectY(u8),
 }
 
@@ -383,9 +383,9 @@ pub enum NOPU {
 	Immediate(u8),
 	ZeroPage(u8),
 	ZeroPageX(u8),
-	Absolute(u8),
-	AbsoluteX(u8),
-	AbsoluteY(u8),
+	Absolute(u16),
+	AbsoluteX(u16),
+	AbsoluteY(u16),
 }
 
 impl Inst {
